requirements.txt
GPIO
Pygame
opencv-python
pyzmq
imutils
numpy